pyproject.toml
[tool.poetry]
authors     = ["Pritam Soni <23050213+pritamsoni-hsr@users.noreply.github.com>"]
description = "xleap's python sdk"
name        = "xleap"
readme      = "README.md"
version     = "0.1.6"

documentation = "http://docs.xleaplabs.com"
homepage      = "https://xleaplabs.com"
repository    = "https://github.com/xleap-ai/xleap"

include  = ["CHANGELOG.md"]
packages = [{ include = "xleap" }]


  [tool.poetry.dependencies]
  attrs           = ">=21.3.0"
  langchain       = "^0.1.7"
  numpy           = "1.23.5"
  openai          = "^1.9.0"
  pandas          = "^1.5.3"
  python          = "^3.11"
  python-dateutil = "^2.8.0"
  rich            = "^13.7.0"
  rouge-score     = "^0.1.2"
  textstat        = "^0.7.3"
  typer           = "^0.9.0"

  [tool.poetry.group.dev.dependencies]
  mkdocs          = "^1.5.3"
  mkdocs-click    = "^0.8.1"
  mkdocs-material = "^9.5.10"
  mkdocstrings    = { extras = ["crystal", "python"], version = "^0.24.0" }

  [tool.poetry.group.additional.dependencies]
  evaluate              = { version = "^0.4.1", optional = true }
  sentence-transformers = { version = "^2.3.1", optional = true }

[build-system]
build-backend = "poetry.core.masonry.api"
requires      = ["poetry-core"]

[tool.ruff]
line-length = 120
select      = ["F", "I"]

[tool.poetry.extras]
all = ["evaluate", "sentence-transformers"]

[tool.black]
extend-exclude = "_client/*"
include = '\.pyi?$'
line-length = 88
target-version = ['py311']
